% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllSeqDists.R
\name{AllSeqDists}
\alias{AllSeqDists}
\title{Calculate all distances between a set of sequences and a fixed sequence.}
\usage{
AllSeqDists(seqs)
}
\arguments{
\item{seqs}{Matrix or data frame of sequences.}
}
\value{
Vector of the distances from the sequences to 1:N.
}
\description{
Used to calculate the sequence Kendall distance distribution in N! space.
}
\author{
Erik Gregory
}
\keyword{Distance}
\keyword{Kendall}

