% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distL2.R
\name{distL2}
\alias{distL2}
\title{Calculates squared Euclidean distances between two different 
sets of vectors.}
\usage{
distL2(r, centers, mc = 0.25)
}
\arguments{
\item{r}{One set of sequences.}

\item{centers}{Another set of sequences.}

\item{mc}{A constant multiplied to the distance. Default to 0.25 such that
output amounts to all Kendall's distances when two vectors equal
two sets of pairwise relative ordering.}
}
\value{
Matrix where output[i, j] represents the squred euclidean distance 
from sequence "i" in "r" to sequence "j" in "seqs", multiplied by mc.
}
\description{
Amounts to all Kendall's distances when two vectors equal
two sets of score-based pairwise relative ordering, 
i.e., KendallInfo.
}
\author{
Yunlong Jiao
}
\keyword{Euclidean}
\keyword{Squared}
\keyword{distance}

