% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distL2mat.R
\name{distL2mat}
\alias{distL2mat}
\title{Convert a Kendall kernel matrix to a Kendall distance matrix}
\usage{
distL2mat(kmat, r, seqs = NULL, type = c("type-b", "type-a"), mc = 0.25)
}
\arguments{
\item{kmat}{Kendall kernel matrix of dimension m1 x m2, correlation type correponding to "type".}

\item{r, seqs}{Matrices of m1,m2 sequences in rows with n ranked items in cols, by default "seq" set equal to "r".}

\item{type}{A character string indicating the type of Kendall correlation.}

\item{mc}{A normalization constant default to 0.25 such that output squared Euclidean distance amounts exactly to Kendall distances.}
}
\value{
A matrix of dimension m1 x m2 where entry [i,j] is the Kendall distance from sequence "i"
in "r" to sequence "j" in "seqs".
}
\description{
Normalized squared Euclidean distances in feature space recovers 
exactly the Kendall distance.
}
\note{
Kernel trick is explored in the sense that "r" and "seq" are only used for checking dimensions and getting attributes but not used explicitly to compute the distance.
This is particularly interesting when data is high-dimensional in constrast to rather few observations (m1,m2>>n).
}
\examples{
r <- lapply(1:20, function(i) sample.int(1000))
r <- do.call('rbind', r)

# Kendall kernel matrix first
kmat <- pcaPP::cor.fk(t(r))
# then Kendall distance matrix
dmat <- distL2mat(kmat, r, type = "type-b")

# notably the above is equivalent to directly calling AllKendall() with kernel trick
dmat <- AllKendall(r, high.dim = TRUE)

}
\references{
Kendall rank correlation coefficient: \url{https://en.wikipedia.org/wiki/Kendall_rank_correlation_coefficient}

Jiao, Y., & Vert, J.-P. (2016). The Kendall and Mallows Kernels for Permutations. 2016. \href{https://hal.archives-ouvertes.fr/hal-01279273}{hal-01279273}
}

