% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupgen.R
\name{dupgen}
\alias{dupgen}
\title{Translate tied observations into compatible full rankings}
\usage{
dupgen(x)
}
\value{
compatible rankings in rows of a matrix if ranking contains duplicated ranks
}
\description{
Ancillary function for duplicating compatible full rankings for partial rankings with tied observations
}

