% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EStep.R
\name{EStep}
\alias{EStep}
\title{The E-step of the EM algorithm}
\usage{
EStep(R, r, p, lambda, G, N, C, all.dists, use.logC = FALSE)
}
\arguments{
\item{R}{List of length "G". Current cluster modal sequences.}

\item{r}{A matrix with "N" sequences of full rankings in rows.}

\item{p}{Numeric vector. The proportion probability of each cluster.}

\item{lambda}{Numeric vector. The lambda parameters from Mallows model for each cluster.}

\item{G}{Number of clusters, equal to length(R).}

\item{N}{Number of rows in the data.}

\item{C}{Vector of normalizing coefficients for the clusters.}

\item{all.dists}{For efficiency, provide all of the Kendall distances
between each sequence and each cluster mode, that is the result of AllKendall(r,do.call('rbind',R)).}

\item{use.logC}{Logical. Whether or not to use log-sum-exp trick.}
}
\value{
Matrix of dimension N x G where output[i,j] represents the current fuzzy assignment membership probability, 
namely the updated probability that subject "i" belongs to cluster "j".
}
\description{
Assigns each ranking the probability that it belongs to each cluster, given
current parameters.
}
\author{
Yunlong Jiao
}
\references{
Murphy, T. B., & Martin, D. (2003). Mixtures of distance-based models for ranking data. Computational statistics & data analysis, 41(3), 645-655.
}
\keyword{expectation}
\keyword{maximization}

