% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianmix.R
\name{gaussianmix}
\alias{gaussianmix}
\title{Fits a Gaussian mixture model to feature points of ranking data.}
\usage{
gaussianmix(datas, G, iter, weights, eqlam, tol = 0.001,
  sigma.control = 0.001, key = "kernelGaussian", verbose = FALSE,
  cov.constr = TRUE, cal.mallowslike = TRUE, mc = 0.25)
}
\arguments{
\item{datas}{Matrix of fully-ranked data or permutations.}

\item{G}{Number of modes, 2 or greater.}

\item{iter}{Maximum number of iterations.}

\item{weights}{Vector of frequencies of each permutation observed.}

\item{tol}{Stopping precision.}

\item{sigma.control}{Control dirac-like distribution for singletons.}

\item{cov.constr}{Logical. Whether or not contrain covariance matrix to be diagonal.}

\item{cal.mallowslike}{Logical. Compute Mallows likelihood wrt permutations.}
}
\value{
See output of FormatOut.
}
\description{
First map rank data to feature space with Kendall embedding,
and then fit a Gaussian mixture in the feature space with an EM algorithm, 
for clustering permutations.
}
\author{
Yunlong Jiao
}
\keyword{Gaussian}
\keyword{cluster}
\keyword{mixture}

