% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_top.R
\name{kendall_top}
\alias{kendall_top}
\title{Kendall kernel for top-k rankings}
\usage{
kendall_top(x, y)
}
\arguments{
\item{x}{Vector. 
If x is numeric, the rank vector converted from x indicate that larger values mean being preferred.
NAs replace unobserved values.}

\item{y}{Same as x.}
}
\value{
Kendall kernel for top-k rankings defined as kendall tau averaged over all compatible full ranking.
}
\description{
Calculates Kendall kernel between top-k rankings in time O (klogk),
where ties (supposed few) are broken by adopting a convolution kernel
averaging compatible rankings without ties.
}
\examples{
x <- c(1.5, 0.1, NA, -4, NA)
y <- c(NA, NA, 0, 3, NA)
kendall_top(x, y)

}
\references{
Jiao, Y., & Vert, J.-P. (2016). The Kendall and Mallows Kernels for Permutations. 2016. \href{https://hal.archives-ouvertes.fr/hal-01279273}{hal-01279273}
}
\author{
Yunlong Jiao
}
\keyword{Kendall}
\keyword{Kernel}
\keyword{Top-kRanking}
