% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_total.R
\name{kendall_total}
\alias{kendall_total}
\title{Kendall kernel for total rankings}
\usage{
kendall_total(x, y = NULL)
}
\arguments{
\item{x}{A vector or a matrix of m1 sequences in rows and orders of n items in cols.
If x is numeric, the rank vector converted from x indicate that larger values mean being preferred.
NAs are not allowed.}

\item{y}{Same as x. By default "y" is set equal to "x".}
}
\value{
Kendall kernel for total rankings defined as type-b (soft version) of kendall kernel.
}
\description{
Calculates Kendall kernel between total rankings in time O (nlogn),
where ties are dealt with type-b (soft version) of kendall kernel
}
\examples{
x <- c(1.5, 0.1, 0, -4, 0)
y <- c(0, 0, 0, 3, 0)
kendall_total(x, y)

}
\references{
Kendall rank correlation coefficient: \url{https://en.wikipedia.org/wiki/Kendall_rank_correlation_coefficient}

Jiao, Y., & Vert, J.-P. (2016). The Kendall and Mallows Kernels for Permutations. 2016. \href{https://hal.archives-ouvertes.fr/hal-01279273}{hal-01279273}
}
\keyword{Kendall}
\keyword{Kernel}
\keyword{TotalRanking}

