% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_weight.R
\name{kendall_weight}
\alias{kendall_weight}
\title{Weighted Kendall kernel for total rankings}
\usage{
kendall_weight(x, y)
}
\arguments{
\item{x, y}{Vector. 
If \code{x} is numeric, the rank vector converted from \code{x} indicate that larger values mean being preferred.
NAs are not allowed.}
}
\value{
Weighted Kendall kernel for total rankings,
where ties (supposed few) are broken by adopting a convolution kernel averaging compatible rankings without ties.
}
\description{
Calculates weighted Kendall kernel between total rankings in time \code{O(nlogn)},
where ties (supposed few) are broken by adopting a convolution kernel
averaging compatible rankings without ties.
}
\examples{
x <- c(1.5, 0.1, 0, -4, 0)
y <- c(0, 0, 0, 3, 0)
kendall_weight(x, y)

}
\references{
Yunlong Jiao, Jean-Philippe Vert. "The Weighted Kendall and High-order Kernels for Permutations." arXiv preprint arXiv:1802.08526, 2018. \href{https://arxiv.org/abs/1802.08526}{arXiv:1802.08526}
}
\author{
Yunlong Jiao
}
\keyword{Kendall}
\keyword{Kernel}
\keyword{TotalRanking}
\keyword{Weighted}
