% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KendallInfo.R
\name{KendallInfo}
\alias{KendallInfo}
\title{Kendall embedding of pairwise relative ordering}
\usage{
KendallInfo(r)
}
\arguments{
\item{r}{A vector or a matrix of dimension N x n with sequences in rows.}
}
\value{
A matrix of dimension N x choose(n,2) with entry values -1/1/0 representing pairwise comparisons of
vector values for each row. Specifically, a -1 value denotes that there is 
an increase between the two columns, 1 a decrease, and 0 indicates that 
the column values are identical in the same row.
}
\description{
Performs between-column comparison on a matrix of sequences denoting 
sign([,i] - [,j]) for i < j.
}
\note{
A matrix with one row is returned if the input "r" is a vector.
}
\examples{
r <- do.call('rbind', combinat::permn(1:5))
KendallInfo(r)

}
\references{
Jiao, Y., & Vert, J.-P. (2016). The Kendall and Mallows Kernels for Permutations. 2016. \href{https://hal.archives-ouvertes.fr/hal-01279273}{hal-01279273}
}
\author{
Yunlong Jiao
}
\keyword{Embedding}
\keyword{Kendall}
