% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsumexp.R
\name{logsumexp}
\alias{logsumexp}
\title{Computing Log-Sum-Exp with a common trick}
\usage{
logsumexp(x, byrow = TRUE, bycol = !byrow)
}
\arguments{
\item{x}{A vector or a matrix of numerics.}

\item{byrow}{Logical. Computes by rows if a matrix "x" is provided.}

\item{bycol}{Logical. Computes by cols if a matrix "x" is provided.}
}
\value{
Log-Sum-Exp of the numbers in the vector "x" or row-/column-wise Log-Sum-Exp of the numbers in matrix "x".
}
\description{
Computes log-sum-exp of a series of (typically large in absolute value) numbers with a trick
}
\examples{
x <- c(-1000, -999, -1000)
logsumexp(x)

}
\author{
Yunlong Jiao
}
\references{
Computing Log-Sum-Exp: \url{https://hips.seas.harvard.edu/blog/2013/01/09/computing-log-sum-exp/}
}

