% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mallows_kernel.R
\name{Mallows_kernel}
\alias{Mallows_kernel}
\title{Fits a kernel version of Mallows mixture model to ranking data.}
\usage{
Mallows_kernel(datas, G, iter, weights, eqlam, tol = 0.001, iterin = iter,
  key = "kernelMallows", logsumexp.trick = TRUE, exh = FALSE)
}
\arguments{
\item{datas}{Matrix of fully-ranked data or permutations.}

\item{G}{Number of modes, 2 or greater.}

\item{iter}{Maximum number of iterations for EM algorithm.}

\item{weights}{Vector of frequencies of each permutation observed.}

\item{eqlam}{Logical. Whether to contrain disperson parameters equal.}

\item{tol}{Stopping precision.}

\item{iterin}{Maximum number of iterations for iterative optimization between center and dispersion.}

\item{logsumexp.trick}{Logical. Use logsumexp trick to compute log-likelihood.}

\item{exh}{Logical. Whether to do iterative optimization between center and dispersion.}
}
\value{
See output of FormatOut.
}
\description{
Fits the Mallows mixture model to full ranking data, using
Kendall distance and an EM algorithm, for clustering permutations, 
where centers are abitrary points in feature space induced by Kendall embedding.
}
\author{
Yunlong Jiao
}
\keyword{Mallows}
\keyword{cluster}
\keyword{kernel}
\keyword{mixture}

