% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextTable.R
\name{NextTable}
\alias{NextTable}
\title{Calculates the table of Kendall distances in (N+1)! space, given those in N!
space.}
\usage{
NextTable(last.table, N.last)
}
\arguments{
\item{last.table}{Table of distances in N! space.}

\item{N.last}{N}
}
\value{
Table of distances in (N+1)! space.
}
\description{
This is identical to counting the number of fully-ordered vectors at each
bubble sort distance in (N+1)! space.
}
\note{
Taken directly from RMallow package https://cran.r-project.org/web/packages/RMallow/index.html
}
\author{
Erik Gregory
}
\keyword{Kendall}
\keyword{bubblesort}

