% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankAggreg.R
\name{RankAggreg}
\alias{RankAggreg}
\title{Common ranking aggregation methods for permutations}
\usage{
RankAggreg(r, z = NULL, infos = NULL, perm = NULL, key = c("borda",
  "copeland", "brute"))
}
\arguments{
\item{r}{A vector or a matrix of sequences in rows.}

\item{z}{A vector of weights/frequencies of observations or a matrix of probability of cluster membership 
for each sequence and each cluster. Set by default a constant vector of 1.}

\item{infos}{The result of calling \code{\link{KendallInfo}}. Optional for speeding up computation.}

\item{perm}{A matrix of full permutations for brute-force search of optimal Kemeny consensus. Only effective when "\code{key}" equals to "\code{brute}".}

\item{key}{A character string indicating the ranking aggregation method to find centers.
\itemize{
\item \code{borda} denotes the Borda count
\item \code{copeland} denotes the Copeland's aggregated ranking
\item \code{brute} denotes the optimal Kemeny consensus found by brute-force search
}}
}
\value{
List of length 1 if "\code{z}" is a vector, or \code{ncol(z)} if "\code{z}" is a matrix, each entry being the modal sequence in each cluster.
}
\description{
Used to update modal sequences of each cluster in the EM algorithm when fitting Mallows mixture models.
}
\examples{
r <- do.call("rbind", list(1:5, 5:1, c(2,4,1,5,3)))
RankAggreg(r, key = "borda") # Borda count for sequences in "r"

}
\seealso{
\code{\link{KendallInfo}}
}
\author{
Yunlong Jiao
}
\keyword{RankAggregation}
\keyword{TotalRanking}
