% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllKendall.R
\name{AllKendall}
\alias{AllKendall}
\title{All Kendall's distances between two sets of total rankings or real-valued vectors}
\usage{
AllKendall(r, seqs = NULL, data.info = NULL, use.kernel.trick = FALSE,
  kmat = NULL, type = c("type-b", "type-a"), mc = 0.25)
}
\arguments{
\item{r}{A vector or a matrix of \code{m1} sequences in rows and orders of \code{n} items in cols.}

\item{seqs}{Another vector or a matrix of \code{m2} sequences in rows and orders of \code{n} items in cols. By default "\code{seqs}" is set equal to "\code{r}".}

\item{data.info}{Optional argument giving the Kendall embedding of "\code{r}", that is the result of calling \code{\link{KendallInfo}}, 
to facilitate computing Kendall's difference for "\code{r}" to "\code{seqs}" without exploring the kernel trick.}

\item{use.kernel.trick}{Logical. Indicator of whether the kernel trick is explored. 
This is particularly interesting when the number of items to be ranked is large \code{(m1, m2 >> n)} 
and will use \code{\link[pcaPP]{cor.fk}} for fast computation. 
By default (set \code{FALSE}), Kendall embedding is explicitly computed; 
otherwise kernel trick is explored.}

\item{kmat}{Kendall kernel matrix of dimension \code{m1 x m2}, correlation type correponding to "\code{type}". If given, kernel trick is explored directly.}

\item{type}{A character string indicating the type of Kendall correlation for "\code{kmat}".}

\item{mc}{A normalization constant default to 0.25 such that output normalized squared Euclidean distance in the feature space induced by Kendall embedding amounts exactly to Kendall distances.}
}
\value{
A matrix of dimension \code{m1 x m2} where entry \code{[i,j]} is the distance from sequence \code{i}
in "\code{r}" to sequence \code{j} in "\code{seqs}".
}
\description{
Calculates all of the Kendall's distances between two sets of total 
rankings or real-valued vectors.
}
\note{
Kernel trick is explored in the sense that "\code{r}" and "\code{seq}" are only used for checking dimensions and getting attributes but not used explicitly to compute the distance.
Option "\code{use.kernel.trick}" set \code{TRUE} or \code{FALSE} may give slightly different results due to computation precision of two implementations.
}
\examples{

#### Ex 1: Compute Kendall distance matrix and Mallows kernel matrix
data1 <- do.call("rbind", list(1:5, 5:1, c(3, 2, 1, 4, 5)))
data2 <- do.call("rbind", list(1:5, 5:1))

# Kendall distance matrix
s.K.d.mat <- AllKendall(data1, data2)

# Mallows kernel matrix with dispersion parameter lambda
lambda <- 0.1
M.k.kmat <- exp(-lambda * s.K.d.mat)

#### Ex 2: Why kernel trick?
r <- lapply(1:20, function(i) sample.int(1000, replace = TRUE))
r <- do.call('rbind', r)
dim(r)

# I) Without kernel trick
pt <- proc.time()
dmat1 <- AllKendall(r, use.kernel.trick = FALSE)
proc.time() - pt

# II) With kernel trick (should be much faster in this setting)
require(pcaPP)
pt <- proc.time()
dmat2 <- AllKendall(r, use.kernel.trick = TRUE)
proc.time() - pt

# NOTE: dmat1 and dmat2 may return slightly different values due to computation precision
isTRUE(all.equal(dmat1, dmat2, check.attributes = FALSE)) # May sometimes output FALSE
isTRUE(max(abs(dmat1 - dmat2)) < 1e-6) # Should always output TRUE

}
\references{
Kendall's tau rank correlation coefficient: \url{https://en.wikipedia.org/wiki/Kendall_rank_correlation_coefficient}

Yunlong Jiao, Jean-Philippe Vert. "The Kendall and Mallows Kernels for Permutations." IEEE Transactions on Pattern Analysis and Machine Intelligence (TPAMI), vol. 40, no. 7, pp. 1755-1769, 2018. \href{https://doi.org/10.1109/TPAMI.2017.2719680}{DOI:10.1109/TPAMI.2017.2719680}
}
\seealso{
\code{\link{KendallInfo}}, \code{\link[pcaPP]{cor.fk}}
}
\author{
Yunlong Jiao
}
\keyword{Distance}
\keyword{Kendall}
\keyword{TotalRanking}
