% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllKendall.R
\name{AllKendall}
\alias{AllKendall}
\title{All Kendall's distances between two sets of total rankings or real-valued vectors}
\usage{
AllKendall(r, seqs = NULL, data.info = NULL, high.dim = FALSE)
}
\arguments{
\item{r}{A vector or a matrix of m1 sequences in rows.}

\item{seqs}{Another vector or a matrix of m2 sequences in rows, by default set to equal to "r".}

\item{data.info}{Optional argument giving the Kendall embedding of "r", that is the result of KendallInfo(r), 
to facilitate computing Kendall's difference for "r" to "seqs".}

\item{high.dim}{Logical indicating whether the number of items to be ranked 
is high. By default (set FALSE), Kendall embedding is explicitly computed; 
otherwise kernel trick is explored.}
}
\value{
A matrix of dimension m1 x m2 where entry [i,j] is the distance from sequence "i"
in "r" to sequence "j" in "seqs".
}
\description{
Calculates all of the Kendall's distances between two sets of total 
rankings or real-valued vectors.
}
\note{
Option "high.dim" set TRUE or FALSE may give slightly different results due to computation precision.
}
\examples{
# Ex 1: compute Kendall distance matrix and Mallows kernel matrix

data1 <- do.call("rbind", list(1:5, 5:1, c(3, 2, 1, 4, 5)))
data2 <- do.call("rbind", list(1:5, 5:1))

# Kendall distance matrix
s.K.d.mat <- AllKendall(data1, data2)

# Mallows kernel matrix with dispersion parameter lambda
lambda <- 0.1
M.k.kmat <- exp(-lambda * s.K.d.mat)

# Ex 2: why kernel trick?

r <- lapply(1:20, function(i) sample.int(1000))
r <- do.call('rbind', r)
dim(r)

# I) without kernel trick
pt <- proc.time()
dmat1 <- AllKendall(r, high.dim = FALSE)
proc.time() - pt

# II) with kernel trick
pt <- proc.time()
dmat2 <- AllKendall(r, high.dim = TRUE)
proc.time() - pt

# dmat1 and dmat2 may be slightly different due to computation precision

}
\author{
Yunlong Jiao
}
\references{
Kendall rank correlation coefficient: \url{https://en.wikipedia.org/wiki/Kendall_rank_correlation_coefficient}

Jiao, Y., & Vert, J. P. (2015). The Kendall and Mallows kernels for permutations. In Proceedings of the 32nd International Conference on Machine Learning (ICML-15) (pp. 1935-1944).
}
\keyword{Kendall}
\keyword{distance}

