% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_partial.R
\name{kendall_partial}
\alias{kendall_partial}
\title{Kendall kernel for interleaving partial rankings}
\usage{
kendall_partial(x, y)
}
\arguments{
\item{x}{Numeric vector. A rank vector is converted from x then,
where larger values indicate being preferred and NA replace unobserved values.}

\item{y}{Same as x.}
}
\value{
Kendall kernel for interleaving partial rankings defined as kendall tau averaged over all compatible full ranking.
}
\description{
Calculates Kendall kernel between interleaving partial rankings in time O (klogk),
where ties (supposed few) are broken by adopting a convolution kernel
averaging compatible rankings without ties.
}
\examples{
x <- c(1.5, 0.1, NA, -4, NA)
y <- c(NA, NA, 0, 3, NA)
kendall_partial(x, y)

}
\author{
Yunlong Jiao
}
\references{
Jiao, Y., & Vert, J. P. (2015). The Kendall and Mallows kernels for permutations. In Proceedings of the 32nd International Conference on Machine Learning (ICML-15) (pp. 1935-1944).
}
\keyword{Kendall}
\keyword{interleaving}
\keyword{kernel}
\keyword{partial}
\keyword{ranking}

