% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_total.R
\name{kendall_total}
\alias{kendall_total}
\title{Kendall kernel for total rankings}
\usage{
kendall_total(x, y)
}
\arguments{
\item{x, y}{Vector.
If \code{x} is numeric, the rank vector converted from \code{x} indicate that larger values mean being preferred.
NAs are not allowed.}
}
\value{
Kendall kernel for total rankings,
where ties are dealt with by type-b soft version of Kendall tau correlation.
}
\description{
Calculates Kendall kernel between total rankings in time \code{O(nlogn)}, 
where ties are dealt with by type-b soft version of Kendall tau correlation.
}
\examples{
x <- c(1.5, 0.1, 0, -4, 0)
y <- c(0, 0, 0, 3, 0)
kendall_total(x, y)

}
\references{
Kendall rank correlation coefficient: \url{https://en.wikipedia.org/wiki/Kendall_rank_correlation_coefficient}

Yunlong Jiao, Jean-Philippe Vert. "The Kendall and Mallows Kernels for Permutations." IEEE Transactions on Pattern Analysis and Machine Intelligence (TPAMI), vol. 40, no. 7, pp. 1755-1769, 2018. \href{https://doi.org/10.1109/TPAMI.2017.2719680}{DOI:10.1109/TPAMI.2017.2719680}
}
\keyword{Kendall}
\keyword{Kernel}
\keyword{TotalRanking}
