% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_total.R
\name{kendall_total}
\alias{kendall_total}
\title{Kendall kernel for total rankings}
\usage{
kendall_total(x, y)
}
\arguments{
\item{x}{Numeric vector. A rank vector is converted from x then,
where larger values indicate being preferred and NA not allowed.}

\item{y}{Same as x.}
}
\value{
Kendall kernel for total rankings defined as type-a (soft version) of kendall kernel.
}
\description{
Calculates Kendall kernel between total rankings in time O (nlogn),
where ties are dealt with type-a (soft version) of kendall kernel
}
\note{
The implementation is directly taken as pcaPP::cor.fk
}
\examples{
## Not run:

x <- c(1.5, 0.1, 0, -4, 0)
y <- c(0, 0, 0, 3, 0)
kendall_total(x, y)

## End(Not run)
}
\keyword{kendall}
\keyword{kernel}
\keyword{ranking}
\keyword{total}

