% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_total.R
\name{kendall_total}
\alias{kendall_total}
\title{Kendall kernel for total rankings}
\usage{
kendall_total(x, y)
}
\arguments{
\item{x}{Numeric vector. A rank vector is converted from x then,
where larger values indicate being preferred and NA not allowed.}

\item{y}{Same as x.}
}
\value{
Kendall kernel for total rankings defined as type-a (soft version) of kendall kernel.
}
\description{
Calculates Kendall kernel between total rankings in time O (nlogn),
where ties are dealt with type-b (soft version) of kendall kernel
}
\note{
The implementation is directly taken from pcaPP::cor.fk()
}
\examples{
x <- c(1.5, 0.1, 0, -4, 0)
y <- c(0, 0, 0, 3, 0)
kendall_total(x, y)

}
\references{
Kendall rank correlation coefficient: \url{https://en.wikipedia.org/wiki/Kendall_rank_correlation_coefficient}

Jiao, Y., & Vert, J. P. (2015). The Kendall and Mallows kernels for permutations. In Proceedings of the 32nd International Conference on Machine Learning (ICML-15) (pp. 1935-1944).
}
\keyword{kendall}
\keyword{kernel}
\keyword{ranking}
\keyword{total}

