% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KendallInfo.R
\name{KendallInfo}
\alias{KendallInfo}
\title{Kendall embedding of pairwise relative ordering}
\usage{
KendallInfo(r)
}
\arguments{
\item{r}{A vector or a matrix of dimension \code{N x n} with sequences in rows.}
}
\value{
A matrix of dimension \code{N x choose(n,2)} with entry values -1/1/0 representing pairwise comparisons of
vector values for each row. Specifically, a -1 value denotes that there is 
an increase between the two columns, 1 a decrease, and 0 indicates that 
the column values are identical in the same row.
}
\description{
Performs between-column comparison on a matrix of sequences denoting 
\code{sign([,i] - [,j])} for \code{i < j}.
}
\note{
A matrix with one row is returned if the input "\code{r}" is a vector.
}
\examples{
r <- do.call('rbind', combinat::permn(1:5))
KendallInfo(r)

}
\references{
Yunlong Jiao, Jean-Philippe Vert. "The Kendall and Mallows Kernels for Permutations." IEEE Transactions on Pattern Analysis and Machine Intelligence (TPAMI), vol. 40, no. 7, pp. 1755-1769, 2018. \href{https://doi.org/10.1109/TPAMI.2017.2719680}{DOI:10.1109/TPAMI.2017.2719680}
}
\author{
Yunlong Jiao
}
\keyword{Embedding}
\keyword{Kendall}
