% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KendallInfo.R
\name{KendallInfo}
\alias{KendallInfo}
\title{Kendall embedding of pairwise relative ordering}
\usage{
KendallInfo(r)
}
\arguments{
\item{r}{Matrix of sequences. Instances in rows.}
}
\value{
Matrix of -1s, 1s, and 0s representing pairwise comparisons of
vector values.
}
\description{
Performs each column-wise comparison on a matrix of sequences.  A -1 value
denotes that there is an increase between the two columns, 1 a decrease, and
0 indicates that the column values are identical in the row.
}
\author{
Yunlong Jiao
}
\references{
Jiao, Yunlong and Vert, Jean-Philippe. The Kendall and Mallows Kernels for Permutations.
Proceedings of The 32nd International Conference on Machine Learning (ICML), JMLR: W&CP 37, 1935-1944, 2015.
}
\keyword{Distance}
\keyword{Kendall}

