% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsumexp.R
\name{logsumexp}
\alias{logsumexp}
\title{Computing Log-Sum-Exp with a common trick}
\usage{
logsumexp(x, byrow = TRUE, bycol = !byrow)
}
\arguments{
\item{x}{A matrix or vector of numerics.}

\item{byrow}{Logical. Computes by rows if a matrix is provided.}

\item{bycol}{Logical. Computes by cols if a matrix is provided.}
}
\value{
Log-Sum-Exp of the numbers in vector or row-/column-wise Log-Sum-Exp of the numbers in matrix.
}
\description{
Computes log-sum-exp of a series of (typically large in absolute value) numbers with a trick
}
\examples{

## Not run:

x <- c(-1000, -999, -1000)
logsumexp(x)

## End(Not run)
}
\author{
Yunlong Jiao
}
\references{
https://hips.seas.harvard.edu/blog/2013/01/09/computing-log-sum-exp/
}

