% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mallows.R
\name{Mallows}
\alias{Mallows}
\title{Fits a Mallows mixture model to ranking data.}
\usage{
Mallows(datas, G, weights = NULL, iter = 100, tol = 0.001,
  logsumexp.trick = TRUE, iterin = iter, exhkey = "_Exh",
  eqlamkey = "_Eqlam", key = c("copelandMallows", "bruteMallows",
  "bordaMallows", "kernelMallows", "kernelMallows_Exh", "copelandMallows_Eqlam",
  "bruteMallows_Eqlam", "bordaMallows_Eqlam", "kernelMallows_Eqlam",
  "kernelMallows_Exh_Eqlam", "kernelGaussian", "kernelGaussian_Eqlam"))
}
\arguments{
\item{datas}{Matrix of fully-ranked data or permutations.}

\item{G}{Number of modes, 2 or greater.}

\item{weights}{Vector of frequencies of each permutation observed.}

\item{iter}{Maximum number of iterations.}

\item{tol}{Stopping precision.}

\item{logsumexp.trick}{Logical. Use logsumexp trick to compute log-likelihood.}
}
\value{
See output of FormatOut.
}
\description{
Fits the Mallows mixture model to full ranking data, using
Kendall distance and an EM algorithm, for clustering permutations.
}
\author{
Yunlong Jiao
}
\references{
"Mixtures of distance-based models for ranking data". Thomas 
Brendan Murphy & Donal Martin. 1 April 2002. Computational Statistics & 
Data Analysis 41 (2003) 645-655.
}
\keyword{Mallows}
\keyword{cluster}
\keyword{mixture}

