% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MallowsCV.R
\name{MallowsCV}
\alias{MallowsCV}
\title{Fits a Mallows mixture model to ranking data.}
\usage{
MallowsCV(datas, G, weights = NULL, ..., seed = 26921332, nfolds = 5,
  nrepeats = 10, ntry = 3, logsumexp.trick = TRUE)
}
\arguments{
\item{datas}{Matrix of fully-ranked data or permutations.}

\item{G}{Number of modes, 2 or greater.}

\item{weights}{Vector of frequencies of each permutation observed,
which cannot contain 0 and should be of equal length with nrow of datas since cv split is done by partitioning weights!}

\item{...}{Arguments passed to main function Mallows.}

\item{nfolds}{nfold-fold cv created each time.}

\item{nrepeats}{cv repeated nrepeats times.}

\item{ntry}{Number of random initializations to restart each run and best fit with max likelihood chosen.}

\item{logsumexp.trick}{Logical. Use logsumexp trick to compute log-likelihood.}
}
\value{
See output of FormatOut.
}
\description{
Fits the Mallows mixture model to full ranking data, using
Kendall distance and an EM algorithm, for clustering permutations.
}
\author{
Yunlong Jiao
}
\keyword{Mallows}
\keyword{cluster}
\keyword{mixture}

