% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateR.R
\name{UpdateR}
\alias{UpdateR}
\title{Common ranking aggregation methods}
\usage{
UpdateR(r, z = NULL, infos = NULL, perm = NULL, key = c("borda",
  "copeland", "brute"))
}
\arguments{
\item{r}{A vector or a matrix of sequences in rows.}

\item{z}{A vector of weights/frequencies of observations or a matrix of probability of cluster membership 
for each sequence and each cluster. Set by default a constant vector of 1.}

\item{infos}{The result of KendallInfo(r). Optional for speeding up computation.}

\item{perm}{A matrix of full permutations for brute-force search of optimal Kemeny consensus. Only effective for "key" equal to "brute".}

\item{key}{A character string indicating the ranking aggregation method to find centers.
\itemize{
\item \emph{borda} denotes the Borda count
\item \emph{copeland} denotes the Copeland's aggregated ranking
\item \emph{brute} denotes the optimal Kemeny consensus found by brute-force search
}}
}
\value{
List of length 1 if "z" is a vector, or ncol(z) if "z" is a matrix, each entry being the modal sequence in each cluster.
}
\description{
Used to update modal sequences of each cluster in the EM algorithm when fitting Mallows mixture models.
}
\examples{
r <- do.call("rbind", list(1:5, 5:1, c(2,4,1,5,3)))
UpdateR(r, key = "borda") # Borda count for sequences in "r"

}
\author{
Yunlong Jiao
}
\keyword{aggregation}
\keyword{rank}

