% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateR.R
\name{UpdateR}
\alias{UpdateR}
\title{Common ranking aggregation methods.}
\usage{
UpdateR(r, z = NULL, infos = NULL, perm = NULL, key = c("borda",
  "copeland", "brute"))
}
\arguments{
\item{r}{Matrix of sequences being clustered. Permutations in rows.}

\item{z}{Vector of weights or matrix of probability of cluster membership 
for each sequence and each cluster.}

\item{infos}{The KendallInfo matrix for "r". Optional for speeding up computation.}

\item{perm}{Full list of permutations for brute-force search of optimal center.}

\item{key}{Character. Ranking aggregation method to define centers.}
}
\value{
List. Consensus ranking or new cluster centers for each cluster.
}
\description{
Used in EM algorithm for fitting Mallows model to update modal sequences 
in each cluster.
}
\examples{
## Not run:

r <- do.call("rbind", list(1:5, 5:1, c(2,4,1,5,3)))
UpdateR(r, key = "borda")

## End(Not run)
}
\author{
Yunlong Jiao
}
\keyword{center}
\keyword{cluster}

